Share widget rejig
